Changed contract for mute method.
Mute takes two arguments, an object and a flag. The previous behavior was
for the mute status of the object to be unconditionally set to flag.  This
is still the case for PARTS.  The behavior has changed for groups.   Mute
now may only turn group muting on, a nil flag does not unmute the group.
There is a new unmute method for unmuting groups.

Motivation:

In a project a set of mute statements often appear above group muting
statements.   With the previous behavior it was not possible to mute
individual parts within a group with out the subsequent call to mute on the
group undoing things.   Now (mute grp nil) is ignored and all member parts
maintain their mute status.   The new unmute statement may be used to
unmute a group.  In practice this is usually not necessary as any group
status is re-established each time a section file is reloaded.